refactor: split parameters into multiple files
This is for a couple of reasons:
- Improve maintainability
- Identify which parameters do not have tests (not sure if we need them?)

The parameters are grouped by concern, which is hopefully easier to reason about. For example `ec2` contains parameters concerning an EC2 based stack:
- instance type
- ami
