+++ src/constructs/core/parameters/s3.ts
@@ -1,0 +1,27 @@
+import { RegexPattern } from "../../../constants";
+import type { GuStack } from "../stack";
+import type { GuNoTypeParameterProps } from "./base";
+import { GuStringParameter } from "./base";
+
+export class GuS3ObjectArnParameter extends GuStringParameter {
+  constructor(scope: GuStack, id: string, props: GuNoTypeParameterProps) {
+    super(scope, id, {
+      ...props,
+      allowedPattern: RegexPattern.S3ARN,
+      constraintDescription:
+        "Must be a valid S3 ARN, see https://docs.aws.amazon.com/AmazonS3/latest/dev/BucketRestrictions.html",
+    });
+  }
+}
+
+export class GuDistributionBucketParameter extends GuStringParameter {
+  public static parameterName = "DistributionBucketName";
+
+  constructor(scope: GuStack, id: string = GuDistributionBucketParameter.parameterName) {
+    super(scope, id, {
+      description: "SSM parameter containing the S3 bucket name holding distribution artifacts",
+      default: "/account/services/artifact.bucket",
+      fromSSM: true,
+    });
+  }
+}
